requirements.txt
torch
transformers
pandas
matplotlib
seaborn
numpy
polars
sqlparse
aws-cdk-lib
constructs
psycopg2-binary
SQLAlchemy
boto3
fastapi
mangum
uvicorn
pgvector
openai
